don't finish the simringer and reject if evolving
If we have 5 endpoints we want to contact but we
don't put them into the simringer immediately but
one of them errors out, the simringer is set to
finished. We Don't want that to happen. If we have
say 2 Webrtc endpoints in a browser but our record
for them is stale, drachtio will return with a 503
DNS error because of the .invalid domain. However,
we're still waiting to put more contact URIs into
the simringer because we're waiting on the
registrations to come in from PUSH started
registrations. So this patch tells simringer if it
should see itself as an evolving simringer, if it
is then you can also pass in the total number of
expected URIs. As one after one errors, we reduce
the count but we don't say the simringer is over
until we have no more URIs in our list to call/wait
for.